Move get_theme_from_slug to utils file
