+++ home/templatetags/theme_tags.py
@@ -1,7 +1,8 @@
 from django import template
 from wagtail.core.models import Site
 
-from home.themes import DEFAULT_THEME, THEMES
+from home.themes import DEFAULT_THEME
+from home.utils import get_theme_from_slug
 
 register = template.Library()
 
@@ -19,16 +20,3 @@ def get_current_theme(context):
         if site.root_page.specific.theme:
             theme = get_theme_from_slug(site.root_page.specific.theme)
     return theme
-
-
-def get_theme_from_slug(value):
-    """
-    Returns the ``Theme`` instance from ``themes.THEMES``
-    with a ``slug`` value matching the provided string.
-    """
-    if not value:
-        return
-    for theme in THEMES:
-        if theme.slug == value:
-            return theme
-    raise ValueError(f"No Theme can be found matching the slug '{value}'.")
